Split big component into separate parts
