+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Calendar from "./Calendar";
+import Calendar from "./components/Calendar";
 
 import "./App.css";
