+++ src/components/Calendar/index.js
@@ -1,166 +1,15 @@
 import React from "react";
 import dateFns from "date-fns";
-import cn from "classnames";
 import SimpleStorage from "react-simple-storage";
-import { Tooltip } from 'react-tippy';
 
-import 'react-tippy/dist/tippy.css'
+import Header from './Header';
+import Week from './Week';
+import Body from './Body';
+import Form from './Form';
+import EventsList from './EventsList';
 
-import getCalendarRowsData from './lib/getCalendarRowsData.js'
-
-
-const Header = (props) => {
-  const HEADER_DATE_FORMAT = "MMMM YYYY";
-
-  const [selectedMonth, selectedYear] = dateFns.format(props.currentMonth, HEADER_DATE_FORMAT).split(' ');
-
-  return (
-    <div className="calendar-header calendar-row">
-      <div className="calendar-left-arrow calendar-col calendar-col-left" onClick={props.prevMonth}>
-        <div className="calendar-icon">
-          chevron_left
-        </div>
-      </div>
-      <div className="calendar-header-date calendar-col calendar-col-center">
-        <span>{selectedMonth}<br />{selectedYear}</span>
-      </div>
-      <div className="calendar-right-arrow calendar-col calendar-col-right" onClick={props.nextMonth}>
-        <div className="calendar-icon">chevron_right</div>
-      </div>
-    </div>
-  );
-}
-
-const Week = (props) => {
-  const dateFormat = "dddd";
-  const days = [];
-
-  const day = new Date();
-  const startDate = dateFns.startOfWeek(day);
-
-  for (let i = 0; i < 7; i++) {
-    days.push(
-      <div className="calendar-col calendar-col-center" key={i}>
-        {dateFns.format(dateFns.addDays(startDate, i), dateFormat).substring(0, 2)}
-      </div>
-    );
-  }
-
-  return <div className="calendar-days calendar-row">{days}</div>;
-}
-
-class Form extends React.Component {
-  componentDidMount() {
-    // because react-tippy treat passed html as prop
-    // and marked input not actually in DOM
-    setTimeout(() => this.hourInput.focus(), 0)
-  }
-
-  render() {
-    const {onFormSubmit, onInputChange, formState} = this.props;
-
-    return (
-      <form className="event-form" onSubmit={onFormSubmit}>
-        <div className="event-form__row">
-          <input
-            className="event-form__row-item"
-            type="text" name="hour"
-            placeholder="Hour"
-            onChange={onInputChange}
-            value={formState.hourText}
-            ref={(input) => {this.hourInput = input}}
-          />
-          <input
-            className="event-form__row-item"
-            type="text" name="minute"
-            placeholder="Minute"
-            onChange={onInputChange}
-            value={formState.minuteText}/>
-        </div>
-
-        <div className="event-form__row">
-          <input
-            className="event-form__row-item"
-            type="text" name="eventName"
-            placeholder="Event Name"
-            onChange={onInputChange}
-            value={formState.eventNameText}
-          />
-        </div>
-
-        <div className="event-form__row">
-          <div className="event-form__save-btn-wrapper">
-            <button
-              type="submit"
-              className={cn("event-form__row-item", "event-form__save-btn", `${!formState.isFormInputValid && 'event-form__save-btn--disabled'}`)}
-              disabled={!formState.isFormInputValid}
-            >
-              Save
-            </button>
-          </div>
-        </div>
-      </form>
-    )
-  }
-}
-
-const Body = ({currentMonth, selectedDate, setTooltipState, isTooltipOpen, onDayClick, enableGrid, children}) => {
-  const rows = getCalendarRowsData(currentMonth, selectedDate);
-
-  const renderCell = (dayMeta) => {
-    return (
-      <div
-        className={cn("calendar-col", "calendar-col-center", "calendar-cell", `${enableGrid && 'calendar-cell--with-grid'}`, ...dayMeta.options)}
-        key={dayMeta.day}
-        onClick={onDayClick(dayMeta.day)}
-      >
-        <Tooltip
-          interactive
-          arrow
-          unmountHTMLWhenHide
-          position="bottom"
-          theme="light"
-          trigger="click"
-          open={isTooltipOpen(dayMeta)}
-          onRequestClose={setTooltipState(false, dayMeta.day)}
-          html={children}
-        >
-          <div
-            className="calendar-cell-inner"
-            onClick={setTooltipState(true, dayMeta.day)}
-          >
-            <div>{dayMeta.formattedDate}</div>
-          </div>
-        </Tooltip>
-      </div>
-    )
-  }
-
-  const rowsMarkup = rows.map((row, i) => {
-    return (
-      <div className={cn("calendar-row", `${enableGrid && 'calendar-row--with-grid'}`)} key={i}>
-        { row.map(dayMeta => renderCell(dayMeta, isTooltipOpen)) }
-      </div>
-    )
-  })
-
-  return <div className="calendar-body">{rowsMarkup}</div>;
-}
-
-const EventsList = ({eventsListState}) => {
-  if (!eventsListState.length) { return null }
-
-  const dateFormatter = (date) => dateFns.format(date, 'D MMMM YYYY HH-mm')
-
-  const eventsListMarkup = eventsListState.map((eventItem, i) =>
-    <li className="calendar-event-list__item" key={i}> { dateFormatter(eventItem.day) }: - { eventItem.eventName } </li>)
-
-  return (
-    <ul className="calendar-col calendar-event-list">
-      { eventsListMarkup }
-    </ul>
-  )
-}
+import './index.css';
+import 'react-tippy/dist/tippy.css';
 
 class Calendar extends React.Component {
   static initialState = {
