+++ src/components/Calendar/index.scss
@@ -1,0 +1,312 @@
+/* globals */
+
+@import url(https://fonts.googleapis.com/css?family=Open+Sans:300,400,700);
+@import url(https://fonts.googleapis.com/icon?family=Material+Icons);
+
+$text-color: #777;
+$text-color-light: #ccc;
+$btn-text-color: #eee;
+$border-color: #eee;
+$bg-color: #f9f9f9;
+$neutral-color: #fff;
+$header-bg-color: linear-gradient(315deg, #43d6a9, #79dfdf);
+$days-bg-color: #e6e9ed;
+$body-bg-color: #f5f7fa;
+$selected-date-bg-color: #43d6a9;
+$save-btn-color: #43d6a9;
+$disabled-btn-color: #f5f7fa;
+$widget-width: 720px;
+$header-height: 150px;
+$cell-height: 50px;
+
+@mixin list-reset {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+/* GRID */
+/* Based on flexboxes */
+
+.calendar-body .calendar-row {
+  border-bottom: 1px solid $body-bg-color;
+  &:last-child {
+    border-bottom: none;
+  }
+}
+
+.calendar-body .calendar-row--with-grid {
+  border-bottom: 1px solid $border-color;
+  &:last-child {
+    border-bottom: none;
+  }
+}
+
+.calendar-row {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+
+  margin: 0;
+  padding: 0;
+  width: 100%;
+}
+
+.calendar-col {
+  flex-grow: 1;
+  flex-basis: 0;
+  max-width: 100%;
+}
+
+// To scale columns approptieately when body compressed
+.calendar-body .calendar-col {
+  flex-grow: 0;
+  flex-basis: calc(100%/7);
+  width: calc(100%/7);
+}
+
+.calendar-col-left {
+  justify-content: flex-start;
+  text-align: left;
+}
+.calendar-col-center {
+  justify-content: center;
+  text-align: center;
+}
+.calendar-col-right {
+  justify-content: flex-end;
+  text-align: right;
+}
+
+/* calendar */
+
+.calendar-container {
+  display: block;
+  max-width: $widget-width;
+  margin: 0 auto;
+  margin-top: 60px;
+}
+
+.calendar {
+  display: block;
+  position: relative;
+  width: 100%;
+  background: $neutral-color;
+  border-radius: 6px;
+  border: 1px solid $border-color;
+  box-shadow: 0 4px 20px 0 rgba(0,0,0,0.08);
+}
+
+.calendar-header {
+  height: $header-height;
+  width: 100%;
+  font-size: 24px;
+  font-weight: 700;
+  padding: 10px 0;
+  border-radius: 6px 6px 0 0;
+  border-bottom: 1px solid $border-color;
+  background: $header-bg-color;
+  color: $neutral-color;
+}
+
+.calendar-days {
+  font-weight: 400;
+  background: $days-bg-color;
+  color: $text-color;
+  font-size: 90%;
+  padding: .75em 0;
+  border-bottom: 1px solid $border-color;
+}
+
+.calendar-body {
+  background: $body-bg-color;
+}
+
+.calendar-cell {
+  position: relative;
+  height: $cell-height;
+  overflow: hidden;
+  cursor: pointer;
+  background: $body-bg-color;
+  border-right: 1px solid $body-bg-color;
+  &:last-child {
+    border-right: none;
+  }
+
+  color: $text-color;
+  transition: 0.25s ease-out;
+  border-radius: 3px;
+
+  display: flex;
+
+  // To configure third-party Tooltip
+  div { width: 100% }
+
+  &-inner {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: $cell-height;
+  }
+
+  &:hover {
+    background: darken($bg-color, 5%);
+    transition: 0.5s ease-out;
+  }
+  // Uncomment if greed required
+  // border-right: 1px solid $border-color;
+  // &:last-child {
+  //   border-right: none;
+  // }
+}
+
+.calendar-cell--with-grid {
+  border-right: 1px solid $border-color;
+  &:last-child {
+    border-right: none;
+  }
+}
+
+
+.calendar-cell--selected, .calendar-cell--current {
+  background-color: $selected-date-bg-color;
+
+  &:hover {
+    background: lighten($selected-date-bg-color, 10%);
+    transition: 0.5s ease-out;
+  }
+}
+
+.calendar-cell--disabled {
+  color: $text-color-light;
+  pointer-events: none;
+}
+
+/* FORM */
+.event-form {
+  display: flex;
+  flex-direction: column;
+}
+
+.event-form__row {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.event-form__row-item {
+  flex-grow: 1;
+  flex-basis: 0;
+  margin: 5px;
+  padding: 5px;
+}
+
+.event-form__save-btn-wrapper {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  margin: 0 auto;
+}
+
+.event-form__save-btn {
+  color: $btn-text-color;
+  background: $save-btn-color;
+  border-color: $save-btn-color;
+  min-width: 70px;
+  flex-grow: 0;
+  padding: 10px;
+  margin: 8px;
+  border-radius: 3px;
+  transition: .25s ease-out;
+
+  &:hover {
+    background-color: lighten($save-btn-color, 10%)
+  }
+
+  &--disabled {
+    color: $text-color;
+    background-color: $disabled-btn-color;
+    border-color: $disabled-btn-color;
+    transition: .25s ease-out;
+
+    &:hover {
+      background-color: $disabled-btn-color;
+    }
+  }
+}
+
+.calendar-controls {
+  @include list-reset;
+  flex-grow: 1;
+}
+
+.flush-state-link, .toggle-grid-link {
+  margin-top: 20px;
+  margin-right: 44px;
+  min-width: 100px;
+  color: blue;
+  cursor: pointer;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.calendar-event-list {
+  @include list-reset;
+  flex-grow: 2;
+
+  margin-top: 20px;
+  margin-left: 44px;
+
+  li {
+    border-bottom: 1px solid $border-color
+  }
+
+  li:nth-child(even) {
+    background-color: $body-bg-color;
+  }
+}
+
+/* ICONS */
+
+.calendar-left-arrow {
+  display: flex;
+  align-items: center;
+  padding-left: 10px;
+}
+.calendar-header-date {
+  display: flex;
+  align-items: center;
+}
+.calendar-right-arrow {
+  display: flex;
+  align-items: center;
+  padding-right: 10px;
+}
+
+.calendar-icon {
+  font-family: 'Material Icons', serif;
+  font-style: normal;
+  display: inline-block;
+  vertical-align: middle;
+  line-height: 1;
+  text-transform: none;
+  letter-spacing: normal;
+  word-wrap: normal;
+  white-space: nowrap;
+  direction: ltr;
+  cursor: pointer;
+  transition: .15s ease-out;
+
+  -webkit-font-smoothing: antialiased;
+  text-rendering: optimizeLegibility;
+  -moz-osx-font-smoothing: grayscale;
+  font-feature-settings: 'liga';
+
+  &:hover {
+    transform: scale(1.75);
+    transition: .25s ease-out;
+    color: $neutral-color;
+  }
+}
